requirements.txt
streamlit
qdrant-client
cohere
langchain
langchain_openai
langchain_experimental
langchainhub
langchain_together
fastembed
sentence_transformers
einops
PyPDF2
pypdf
langchain-qdrant
pymongo
certifi
matplotlib
Pillow
